+++ app/src/docs/_examples/richTextView/Basic.example.tsx
@@ -1,11 +1,6 @@
 import React from 'react';
-import {
-    Anchor, ControlWrapper, FlexRow, LabeledInput, LinkButton, RichTextView, TextInput,
-} from '@epam/promo';
-import { Svg } from '@epam/uui-components';
-import { cx } from '@epam/uui-core';
+import { Anchor, FlexRow, IconContainer, LabeledInput, LinkButton, Panel, RichTextView, TextInput } from '@epam/uui';
 import { ReactComponent as Calendar } from '@epam/assets/icons/common/action-calendar-18.svg';
-import basicExampleCss from './BasicExample.module.scss';
 
 export default function BasicExample() {
     return (
@@ -54,7 +49,7 @@ export default function BasicExample() {
                 <Anchor href="/">
                     Click me
                     {' '}
-                    <Svg svg={ Calendar }></Svg>
+                    <IconContainer style={ { fill: 'var(--uui-icon)' } } icon={ Calendar }></IconContainer>
                 </Anchor>
                 {' '}
                 It is those feelings that drive our love of astronomy and our desire to learn more and more about it.
@@ -148,7 +143,7 @@ export default function BasicExample() {
                         <>
                             Demo Component
                             <Anchor href="/">
-                                <Svg svg={ Calendar }></Svg>
+                                <IconContainer style={ { fill: 'var(--uui-icon)' } } icon={ Calendar }></IconContainer>
                             </Anchor>
                             {' '}
                             :
@@ -156,9 +151,7 @@ export default function BasicExample() {
                         </>
                     }
                 >
-                    <ControlWrapper size="36">
-                        <TextInput value="" onValueChange={ () => {} }></TextInput>
-                    </ControlWrapper>
+                    <TextInput value="" onValueChange={ () => {} }></TextInput>
                     If you have no goals to choose, add them on
                     {' '}
                     <Anchor href="/">Home</Anchor>
@@ -172,20 +165,22 @@ export default function BasicExample() {
                 <LinkButton caption="ADD"></LinkButton>
             </FlexRow>
             <h4>{'<Pre>'}</h4>
-            <pre>
-                {`import React from 'react';
-import css from './MyComponent.module.scss';
+            <Panel style={ { background: 'var(--uui-neutral-40)' } }>
+                <pre>
+                    {`import React from 'react';
+import css from './MyComponent.module.css';
 
 export const MyComponent = <div className={ css.myHeader }>`}
-            </pre>
+                </pre>
+            </Panel>
             <h4>Text with size 16</h4>
-            <p className={ cx(basicExampleCss.typography16) }>
+            <p className="uui-typography-16">
                 Lorem ipsum dolor sit amet,
                 {' '}
                 <Anchor href="/">
                     click me
                     {' '}
-                    <Svg svg={ Calendar }></Svg>
+                    <IconContainer style={ { fill: 'var(--uui-icon)' } } icon={ Calendar }></IconContainer>
                 </Anchor>
                 {' '}
                 consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation
@@ -195,13 +190,13 @@ export const MyComponent = <div className={ css.myHeader }>`}
                 <code>{'import { Button } from \'@epam/loveship\''}</code>
                 Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
             </p>
-            <p className={ cx(basicExampleCss.typography16) }>
+            <p className="uui-typography-16">
                 Lorem ipsum dolor sit amet,
                 {' '}
                 <Anchor href="/">
                     click me
                     {' '}
-                    <Svg svg={ Calendar }></Svg>
+                    <IconContainer style={ { fill: 'var(--uui-icon)' } } icon={ Calendar }></IconContainer>
                 </Anchor>
                 {' '}
                 consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation
@@ -212,13 +207,13 @@ export const MyComponent = <div className={ css.myHeader }>`}
                 Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
             </p>
             <h4>Text with size 12</h4>
-            <p className={ cx(basicExampleCss.typography12) }>
+            <p className="uui-typography-12">
                 Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
                 {' '}
                 <Anchor href="/">
                     click me
                     {' '}
-                    <Svg svg={ Calendar }></Svg>
+                    <IconContainer style={ { fill: 'var(--uui-icon)' } } icon={ Calendar }></IconContainer>
                 </Anchor>
                 {' '}
                 quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure reprehenderit in voluptate velit esse cillum dolor
@@ -227,13 +222,13 @@ export const MyComponent = <div className={ css.myHeader }>`}
                 <code>{'import { Button } from \'@epam/loveship\''}</code>
                 Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
             </p>
-            <p className={ cx(basicExampleCss.typography12) }>
+            <p className="uui-typography-12">
                 Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
                 {' '}
                 <Anchor href="/">
                     click me
                     {' '}
-                    <Svg svg={ Calendar }></Svg>
+                    <IconContainer style={ { fill: 'var(--uui-icon)' } } icon={ Calendar }></IconContainer>
                 </Anchor>
                 {' '}
                 quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure reprehenderit in voluptate velit esse cillum dolor