+++ uui-editor/src/SlateEditor.tsx
@@ -1,6 +1,6 @@
-import { ScrollBars } from '@epam/uui-components';
-import { IEditable, IHasCX, IHasRawProps, cx, useForceUpdate, uuiMod } from '@epam/uui-core';
 import React, { Fragment, useMemo, useRef, useState } from 'react';
+import { IEditable, IHasCX, IHasRawProps, cx, useForceUpdate, uuiMod } from '@epam/uui-core';
+import { ScrollBars } from '@epam/uui';
 
 import {
     Plate,
@@ -12,7 +12,6 @@ import {
     usePlateEditorState,
 } from '@udecode/plate-common';
 
-import css from './SlateEditor.module.scss';
 import { createPlateUI } from './components';
 import { migrateSchema } from './migration';
 import { baseMarksPlugin } from './plugins';
@@ -20,6 +19,8 @@ import { MainToolbar, MarksToolbar } from './plugins/Toolbars';
 import { EditorValue } from './types';
 import { defaultPlugins } from './defaultPlugins';
 import { isEditorValueEmpty } from './helpers';
+
+import css from './SlateEditor.module.scss';
 
 const basePlugins: any = [
     baseMarksPlugin(),
@@ -83,21 +84,21 @@ function Editor(props: PlateEditorProps) {
     return (
         <div
             className={ cx(
+                'uui-typography',
                 props.cx,
                 css.container,
                 css['mode-' + (props.mode || 'form')],
                 (!props.isReadonly && isFocused) && uuiMod.focus,
                 props.isReadonly && uuiMod.readonly,
                 props.scrollbars && css.withScrollbars,
-                css.typographyPromo,
-                props.fontSize === '16' ? css.typography16 : css.typography14,
+                props.fontSize === '16' ? 'uui-typography-size-16' : 'uui-typography-size-14',
             ) }
             style={ { minHeight: props.minHeight || 350 } }
             { ...props.rawProps }
         >
             { props.scrollbars
                 ? (
-                    <ScrollBars cx={ css.scrollbars } style={ { width: '100%' } }>
+                    <ScrollBars cx={ css.scrollbars }>
                         { renderEditor() }
                     </ScrollBars>
                 )