+++ changelog.md
@@ -13,8 +13,14 @@
     Any filter/search/sorting change doesn't trigger placeholders' rendering. Old data is shown until new data is received. When reloading is started,
   * `getListProps` now returns `isReloading` flag, which signals that data is reloading.
 * [DataTable][Pickers]: Added Blocker overlay while loading. It appears on initial render and filter, search, sort, page number, and size change.
-* [InputAddon]: added InputAddon component that can be used in cases when you need to add prefix or suffix to the your component with ControlGroup.
+* [InputAddon]: added InputAddon component that can be used in cases when you need to add prefix or suffix to the component with ControlGroup.
 * Use Vitest instead of Jest in Vite UUI template app
+* [RichTextEditor]: added css variables for theming instead `SkinContext`
+* [skinContext]: removed from `UuiContexts`
+* [Button]: added `neutral` value to `color` prop
+* [ScrollBars]: removed 'theme' prop for "Promo" & "Loveship" skins
+* [Checkbox]: removed 'theme' prop for "Loveship" skin
+* typography files removed from skins and imported by default via import into the `@epam/uui` package, if you used mixins before, add the `uui-typography` class where you need it now
 
 **What's Fixed**
 * [UUI Contexts]: fixed context initialization for react 18 with strict mode